fix: observe the final-norm convention independently of the process default dtype
`norm_convention.observe` built its seeded probe with `torch.randn`, which
takes the process-wide default dtype. The model loader's library sets that
default to the load dtype for the duration of every `from_pretrained`, so a
load running in another test's worker thread turned the probe bf16, a
`type_as` fake norm returned bf16, and the offset fold missed the 1e-3
agreement by a half-ulp (relative error 0.00136; seen once in a full suite
run on 2026-09-06, absent in isolation and in the rerun). Pin the probe to
float32 and cast it before the module copy runs, so the observation depends
on nothing outside the module.

The regression test reproduces the mechanism deterministically by setting
the default dtype to bf16 around the observer; it fails against the old
code and passes now. No recorded convention changes: every real norm was
observed in float32 already.

+++ Server/steerlab_server/jlens/norm_convention.py
@@ -138,12 +138,18 @@ def observe(norm_module) -> dict:
     eps = eps_of(norm_module)
 
     generator = torch.Generator().manual_seed(_PROBE_SEED)
-    x = torch.randn(w.numel(), generator=generator) * 3.0
+    # Pin the probe's dtype. ``torch.randn`` otherwise takes the PROCESS default,
+    # and the model loader's library flips that default to the load dtype for
+    # the duration of every ``from_pretrained`` — so a load running in another
+    # thread turned this probe bf16 and a fake norm's ``type_as`` output missed
+    # the fold by a half-ulp (2026-09-06, seen once in a full suite run). The
+    # observation must not depend on process-wide state.
+    x = torch.randn(w.numel(), generator=generator, dtype=torch.float32) * 3.0
     try:
         probe = copy.deepcopy(norm_module).to(device="cpu",
                                               dtype=torch.float32)
         with torch.no_grad():
-            out = probe(x).detach().to(torch.float32).reshape(-1)
+            out = probe(x.to(torch.float32)).detach().to(torch.float32).reshape(-1)
     except Exception as exc:  # noqa: BLE001 — remapped to an actionable error
         raise JLensError(
             f"could not run a float32 copy of {name} to observe its "
